rescan: allow break out of GetUtxo
This commit replaces the use of WaitGroups surrounding
the Enqueueing of a GetUtxoRequest with channels. This
allows the caller to break out if the chain service is
being shutdown, without needing to wait for the rescan
to call Result.

We use a single channel with a buffer size of 1 to
deliver the SpendReport and error back to the invoking
context. If multiple calls to Result occur, we drop
them using a non-blocking select.