sync_test: return reference to rescan object
This commit fixes a bug that could lead to
undefined behavior of the rescans mutex. Previously,
the rescan object was copied-by-value, which included
a mutex as a field. This not allowed, since copying a
mutex inherently returns a *different* mutex. Returning
a reference ensures that the original mutex is
propagated throughout the test.